Rename kine package to k8s_dqlite
Note that go package names are not allowed to contain dashes,
so we'll use an underscore.
