+++ pkg/k8s_dqlite/client/client.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/canonical/k8s-dqlite/pkg/kine/endpoint"
+	"github.com/canonical/k8s-dqlite/pkg/k8s_dqlite/endpoint"
 	clientv3 "go.etcd.io/etcd/client/v3"
 )
