+++ pkg/k8s_dqlite/drivers/sqlite/metrics.go
