+++ pkg/k8s_dqlite/drivers/sqlite/schema.go
