+++ pkg/k8s_dqlite/server/compact.go
