+++ pkg/k8s_dqlite/server/delete.go
