+++ pkg/k8s_dqlite/server/otel.go
