+++ test/util_test.go
@@ -14,11 +14,11 @@ import (
 	"github.com/canonical/go-dqlite/v3/app"
 	"github.com/canonical/k8s-dqlite/pkg/database"
 	"github.com/canonical/k8s-dqlite/pkg/instrument"
-	"github.com/canonical/k8s-dqlite/pkg/kine/drivers/dqlite"
-	"github.com/canonical/k8s-dqlite/pkg/kine/drivers/sqlite"
-	"github.com/canonical/k8s-dqlite/pkg/kine/endpoint"
-	"github.com/canonical/k8s-dqlite/pkg/kine/server"
-	"github.com/canonical/k8s-dqlite/pkg/kine/sqllog"
+	"github.com/canonical/k8s-dqlite/pkg/k8s_dqlite/drivers/dqlite"
+	"github.com/canonical/k8s-dqlite/pkg/k8s_dqlite/drivers/sqlite"
+	"github.com/canonical/k8s-dqlite/pkg/k8s_dqlite/endpoint"
+	"github.com/canonical/k8s-dqlite/pkg/k8s_dqlite/server"
+	"github.com/canonical/k8s-dqlite/pkg/k8s_dqlite/sqllog"
 	"github.com/sirupsen/logrus"
 	clientv3 "go.etcd.io/etcd/client/v3"
 )
